Refactor TUI code
