+++ src/tui/mod.rs
@@ -1,0 +1,6 @@
+mod app;
+mod render;
+mod ui;
+
+pub use app::App;
+pub use ui::ui;
